% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\alias{distribution<-}
\title{define a distribution over greta arrays}
\usage{
distribution(greta_array) <- value

distribution(greta_array)
}
\arguments{
\item{greta_array}{a greta array. For the assignment method it must not
already have a probability distribution assigned.}

\item{value}{a greta array with a distribution (see
\code{\link{distributions}})}
}
\description{
\code{distribution} defines probability distributions over greta
arrays, e.g. to set a model likelihood, or a prior. the \code{description}
function provided by greta.funprior differs from the version in greta in
that it allows you to define functions over (almost) any greta array, be it
a data, variable or operation greta array. The only exceptions are greta
arrays that already have a distribution (there can be only one) and greta
arrays created with discrete operations (we can't compute a gradient).
}
\details{
To enable users to set distributions on variables and operations
whilst ensuring a valid model, greta.funprior computes the appropriate log
Jacobian determinant adjustment by automatic differentiation and adds this
to the model definition for you.

The extract method returns the greta array if it has a distribution,
or \code{NULL} if it doesn't. It has no real use-case, but is included for
completeness
}
\examples{
\dontrun{

# define a model likelihood

# observed data and mean parameter to be estimated
# (explicitly coerce data to a greta array so we can refer to it later)
y <- as_data(rnorm(5, 0, 3))

mu <- uniform(-3, 3)

# define the distribution over y (the model likelihood)
distribution(y) <- normal(mu, 1)

# get the distribution over y
distribution(y)
}
}
