% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funprior.R
\name{funprior}
\alias{funprior}
\title{construct a prior over a parametric function in greta}
\usage{
funprior(..., code)
}
\arguments{
\item{\dots}{greta arrays representing the parameters in your function}

\item{expr}{an expression (R code in curly braces) stating your prior beliefs
about the function. This should use \code{distribution()} on quantities
computed from your function and should no introduce any more parameters to
the model.}
}
\description{
construct a prior over a parametric function in greta
}
\examples{
# growth function
growth <- function (age, alpha, beta, gamma) {
  age2 <- (age - 5) ^ 2
  (alpha * age2) / (beta + age2) + gamma
}

# true parameters
alpha_true <- 3
beta_true <- 5
gamma_true <- 170
sd_true <- 3

# fake data
n <- 100
ages <- runif(n, 0, 50)
mean_heights_true <- growth(ages, alpha_true, beta_true, gamma_true)
heights <- rnorm(n, mean_heights_true, sd_true)

# growth model parameters - improper flat priors on these
alpha <- variable()
beta <- variable(lower = 0)
gamma <- variable(lower = 0)

# the likelihood
sd <- cauchy(0, 1, truncation = c(0, Inf))
mean_heights <- growth(ages, alpha, beta, gamma)
distribution(heights) <- normal(mean_heights, sd)

# we have prior knowledge that the population mean height at age 25 is around
# 171, with a standard deviation (estimation error) of 0.5. We also have prior
# knowledge that the maximum year-on-year growth is around 30cm, with standard
# deviation 5, but must be positive. We provide some code specifying this prior
# belief
fpri <- funprior(alpha, beta, gamma,
                 code = {
                   time_diff <- 1
                   n <- 30
                   time_seq <- seq(1, n, by = time_diff)
                   height_seq <- growth(time_seq, alpha, beta, gamma)
                   height_diffs <- height_seq[-1] - height_seq[-n]
                   grad <- height_diffs / time_diff

                   max_grad <- max(grad)
                   height_25 <- growth(25, alpha, beta, gamma)

                   distribution(max_grad) <-  normal(13, 5, truncation = c(0, Inf))
                   distribution(height_25) <- normal(160, 0.5)
                 })

# the model will have been updated with this prior knowledge, and you can
# inspect the object `fpri` to get intuition about the prior implied on the
# parameters

m <- model(alpha, beta, gamma, sd)
}
