% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{greta.funprior}
\alias{greta.funprior}
\alias{greta.funprior-package}
\title{Specifying Priors Over Parametric Functions in greta}
\description{
It's common in Bayesian statistical analyses to have prior
information on summary statistics of a model rather than on the model
parameters themselves. For example, you may have prior information about
the probability predicted by a logistic regression under some set of
covariates, or the timing of peaks in a dynamical model. In these
situations it can be difficult to determine the priors on the model
parameters that correspond to these priors on summary statistics.
greta.funprior lets you define priors on these summary statistics by
considering them as priors over parametric functions, and provides a simple
interface for fitting models.
}
\examples{

# add a simple example here to introduce the package!

}
